chore(list-box): removed next folder and updated imports
